Extract unused metrics removal into util function
